integrated_projects/Autonomous_Zenith_Optimizer/BACKUP_requirements.freeze.txt
accelerate==1.11.0
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altgraph==0.17.4
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
autograd==1.8.0
autoray==0.8.0
babel==2.17.0
beautifulsoup4==4.14.2
black==25.11.0
bleach==6.3.0
blinker==1.9.0
boto3==1.40.69
botocore==1.40.69
cachetools==6.2.1
ccxt==4.5.19
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cmake==4.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
diastatic-malt==2.15.2
dill==0.4.0
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.121.0
fastjsonschema==2.21.2
filelock==3.20.0
flake8==7.3.0
Flask==3.1.2
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
gast==0.6.0
google==3.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mistune==3.1.4
mpmath==1.3.0
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
narwhals==2.11.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.3.4
openai==2.7.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
paypalrestsdk==1.13.3
pefile==2023.2.7
pennylane==0.43.1
pennylane_lightning==0.43.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.4.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
pure_eval==0.2.3
pycares==4.11.0
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyOpenSSL==25.3.0
pyparsing==3.2.5
PyPDF2==3.0.1
pyserial==3.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytokens==0.3.0
pytz==2025.2
pywin32-ctypes==0.2.3
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
qiskit==2.2.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
rustworkx==0.17.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
scipy-openblas32==0.3.30.0.7
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
starlette==0.49.3
stevedore==5.5.0
streamlit==1.51.0
stripe==13.2.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
trimesh==4.9.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.15
yarl==1.22.0
